Stripped HTML where possible
Removed HTML and converted to 99% markdown.

The only remaining HTML are a few div tags to identify sections for the
accordion and the small javascript at the beginning for implementation.
Note that the opening div tags *must* have markdown="1" as an attribute
for the engine to process the block code as markdown.